+++ src/Symfony/Component/Security/Http/Authenticator/HttpBasicAuthenticator.php
@@ -24,6 +24,7 @@ use Symfony\Component\Security\Http\Authenticator\Passport\Badge\UserBadge;
 use Symfony\Component\Security\Http\Authenticator\Passport\Credentials\PasswordCredentials;
 use Symfony\Component\Security\Http\Authenticator\Passport\Passport;
 use Symfony\Component\Security\Http\EntryPoint\AuthenticationEntryPointInterface;
+use Symfony\Component\Security\Http\SecurityRequestAttributes;
 
 /**
  * @author Wouter de Jong <wouter@wouterj.nl>
@@ -51,7 +52,13 @@ class HttpBasicAuthenticator implements AuthenticatorInterface, AuthenticationEn
 
     public function supports(Request $request): ?bool
     {
-        return $request->headers->has('PHP_AUTH_USER');
+        if (!$request->headers->has('PHP_AUTH_USER')) {
+            $request->attributes->get(SecurityRequestAttributes::UNSUPPORTED_REASONS)?->add('the request has no HTTP basic credentials, as "PHP_AUTH_USER" is not set');
+
+            return false;
+        }
+
+        return true;
     }
 
     public function authenticate(Request $request): Passport