+++ src/Symfony/Bundle/SecurityBundle/CHANGELOG.md
@@ -24,6 +24,7 @@ CHANGELOG
  * Add `user_data_source` and `user_identifier_claim` options to the `oidc_login` authenticator
  * Add `enable_end_session` and `post_logout_redirect_path` options to the `oidc_login` authenticator for RP-Initiated Logout
  * Allow passing a null user to `Security::isGrantedForUser()` and `Security::getAccessDecisionForUser()` to check guest permissions
+ * Show why an authenticator did not support the request in the security profiler panel
 
 8.1
 ---